### 2019-03-23 #1346 @arfon
@mviereck - many thanks for your submission. Please review the submission documentation https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography and format your paper metadata _exactly_ as you find it in the example.
