+++ README.md
@@ -248,7 +248,7 @@ Core concept is:
    - These restrictions can be disabled with x11docker option `--cap-default` or reduced with `--sudouser` or `--user=root`.
    
 That being said, Docker's default capabilities and its seccomp profile are not bad. 
-I am not aware of an escape from a container without an additional isolation degrading configuration.
+I am not aware of an escape from a container without an additional isolation degrading option or configuration.
 However, x11docker follows the [principle of least privilege](https://en.wikipedia.org/wiki/Principle_of_least_privilege). 
 Docker containers should not have capabilities or privileges that they don't need for their job.
 
