### 2019-03-23 6ae3722
update #92 fixes in header
