+++ paper.md
@@ -9,45 +9,70 @@ tags:
   - reproducibility
   - reproducible research
   - prototyping
-author:
+authors:
   - name: Martin Viereck
     orcid: 0000-0002-4532-4020
-date: 13 December 2018
+    affiliation: "1"
+affiliations:
+ - name: Martin Viereck, Germany
+   index: 1
+date: 22 March 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-[`x11docker`](https://github.com/mviereck/x11docker) allows to run graphical applications in a [GNU/Linux container](https://en.wikipedia.org/wiki/Operating-system-level_virtualization) using [Docker](https://en.wikipedia.org/wiki/Docker_(software)).
+[`x11docker`](https://github.com/mviereck/x11docker) allows to run graphical applications in a 
+[GNU/Linux container](https://en.wikipedia.org/wiki/Operating-system-level_virtualization) using 
+[Docker](https://en.wikipedia.org/wiki/Docker_(software)).
 
-Containerisation in general has proven as a useful technology for packaging applications and their dependencies for deployment in cloud-based infrastructures.
-A container is similar in usage to a [virtual machine](https://en.wikipedia.org/wiki/Virtual_machine), but needs less resources. The technical concept, however, is completly different.
+Containerisation in general has proven as a useful technology for packaging applications and their 
+dependencies for deployment in cloud-based infrastructures.
+A container is similar in usage to a [virtual machine](https://en.wikipedia.org/wiki/Virtual_machine), 
+but needs less resources. The technical concept, however, is completly different.
 
-The properties of containers such as portability, dependency packaging, reduced requirements of system environment (i.e. only the container runtime), isolation, 
-and version management of complete application stacks make it a promising candidate to increase computational reproducibility and reusability of research analyses [@boettiger_introduction_2015].
-Their use has been demonstrated in various disciplines, such as software engineering research [@cito_using_2016], bioinformatics [@hosny_algorun_2016], and archeology [@marwick_computational_2017], 
+The properties of containers such as portability, dependency packaging, reduced requirements of 
+system environment (i.e. only the container runtime), isolation, and version management of complete 
+application stacks make it a promising candidate to increase computational reproducibility and 
+reusability of research analyses [@boettiger_introduction_2015].
+Their use has been demonstrated in various disciplines, such as software engineering research 
+[@cito_using_2016], bioinformatics [@hosny_algorun_2016], and archeology [@marwick_computational_2017], 
 and their preservation is an active field of research [@Kratzke_Heuveline_2017,@emsley_framework_2018].
 
-Software and required libraries can be installed in a Docker image to run software that is difficult to install otherwise. It is possible to run outdated versions, specific versions, or latest development code side by side.
+Software and required libraries can be installed in a Docker image to run software that is difficult 
+to install otherwise. It is possible to run outdated versions, specific versions, or latest development 
+code side by side.
 
-The most popular Linux container frontend, Docker, does not provide a [display server](https://en.wikipedia.org/wiki/Display_server) that would allow to run 
-applications with a [graphical user interface](https://en.wikipedia.org/wiki/Graphical_user_interface) (GUI), because Docker is originally built for server software.
+The most popular Linux container frontend, Docker, does not provide a 
+[display server](https://en.wikipedia.org/wiki/Display_server) that would allow to run applications 
+with a [graphical user interface](https://en.wikipedia.org/wiki/Graphical_user_interface) (GUI), 
+because Docker is originally built for server software.
 `x11docker` fills this gap.
 
-`x11docker` allows to execute [Desktop](https://en.wikipedia.org/wiki/Desktop_environment) GUI applications in an isolated environment by running an 
[…]
