### 2019-03-23 #1346 @mviereck
@arfon Thank you for looking at this!
I have reformated the header of `paper.md`, especially `authors` and `affiliations`.
