+++ paper.bib
@@ -57,7 +57,7 @@
   title = {Binder 2.0 - Reproducible, interactive, sharable environments for science at scale},
   author = {Project Jupyter et al.},
   year = {2018},
-  booktitle = {Proceedings of the 17th Python in Science Conference}
+  booktitle = {Proceedings of the 17th Python in Science Conference},
   url = {http://conference.scipy.org/proceedings/scipy2018/pdfs/project_jupyter.pdf},
   doi = {10.25080/Majora-4af1f417-011}
 }
