### 2019-03-23 #1346 @mviereck
Hello @yxliang01 !

Much thanks that you want to look at this.
The PDF generation succeeded: https://github.com/openjournals/joss-papers/blob/joss.01346/joss.01346/10.21105.joss.01346.pdf
However, here the PDF is readable in browser only, it does not open locally if I download it.
