### 2019-03-24 #1346 @arfon
@yxliang01 - thanks for agreeing to review this submission. As @mviereck mentioned, the paper is available [here](https://github.com/openjournals/joss-papers/blob/joss.01346/joss.01346/10.21105.joss.01346.pdf). You can view it in the browser, or download it to open locally (which works fine for me).

Before we get started with the review though, I wanted to do a few editorial checks so please don't start your review until I've started the main `REVIEW` issue.
