### 2019-03-24 59d376f
`--tty`: Workaround: Set environment variables `LINES` and `COLUMNS` to current terminal size
