### 2019-03-25 #1346 @mviereck
@1138-4EB Thanks for your suggestions!

> "completly". typo, should be "completely".

fixed.

> Citation reference @Kratzke_Heuveline_2017 is missing in paper.bib, so the cite is shown as ???? in the PDF.

fixed, changed to `@rechert_preserving_2017`. Maybe @nuest can confirm that he means this one.

> "reduced requirements of system environment (i.e. only the container runtime)". This might be only partially true.

I think this is still true. The script you've linked mainly checks for kernel configuration. My system misses e.g. `apparmor` and `zfs`, but they are not essential for Docker to work.
Core x11docker functionality depends on tools that are quite common.

> "by running an X display server on the host system". This is the default. But, is it a requirement? Isn't it possible to use x11docker in order to run an X server on a machine, containers on a different one and use a third client? Even though it is not the standard use case, it is important not to constrain what x11docker is capable of, or the contexts where it can be a useful complement.

x11docker can also run with e.g. `ssh -X`, and probably some more complex setups are possible. But I think the description should not be too complicated.

> "for deployment in cloud-based infrastructures" might be replaced with "multiple infraestructures" or "heterogeneus infrastructures", since it is certainly not limited to the cloud.
    Below: "because Docker is originally built for server software". It might be adequate to add the reference to cloud-based infraestructures here, instead of the first paragraph.

I am not sure about this. Still important, keep it simple. Sentences that are correct and complete in every detail often get too complicated.

-----------------

Some points I am still not sure about:
 - Overall it sounds a bit complex. Maybe some points could be written more simple. I am not sure about this.
 - @nuest suggested a reference to GUIDock here: https://github.com/mviereck/x11docker/issues/92#issuecomment-459646507
  GUIDock has a similar target as x11docker, but has a quite unpolished code base. I have hardly criticized GUIDock here: https://github.com/mviereck/x11docker/issues/92#issuecomment-460214545
 Maybe it must be added because JOSS requires to cite similar projects.
 - The first half of the summary is about Docker in general, the second part about `x11docker` and what it adds in usability/functionality. Maybe this should be divided with subtitles.

Further suggestions for changes of `paper.md` are appreciated.
