### 2019-03-25 #1346 @1138-4EB
> > "reduced requirements of system environment (i.e. only the container runtime)". This might be only partially true.
>
> I think this is still true. The script you've linked mainly checks for kernel configuration. My system misses e.g. `apparmor` and `zfs`, but they are not essential for Docker to work.
> Core x11docker functionality depends on tools that are quite common.

The point is that some GUI apps can be executed on some systems where docker cannot be installed. In this context, docker does not reduce the requirements, but rather increases them. So, this is not an issue with x11docker itself, but with that fact that running the docker daemon is not trivial. **I think it can be solved just by replacing 'reduced' with 'minimal'**, e.g.: *'(...) dependency packaging, and minimal requirements of system environment (besides the container runtime), isolation (...)'*.

> x11docker can also run with e.g. `ssh -X`, and probably some more complex setups are possible. But I think the description should not be too complicated.

In this case **I'm proposing to remove a limiting addition to the sentence, not to add explicit cases**: '*x11docker allows to execute Desktop GUI applications in an isolated environment by running an X display server <strike>on the host system</strike> and providing it to applications in Docker containers*'.

> > "for deployment in cloud-based infrastructures" might be replaced with "multiple infraestructures" or "heterogeneus infrastructures", since it is certainly not limited to the cloud.
> > Below: "because Docker is originally built for server software". It might be adequate to add the reference to cloud-based infraestructures here, instead of the first paragraph.
>
> I am not sure about this. Still important, keep it simple. Sentences that are correct and complete in every detail often get too complicated.

My main point here is that x11docker is not specially useful in the cloud-based infraestructures where docker is so widespread, but it is a great tool for other not-so-conventional uses of docker, i.e. running desktop applications that use GPU acceleration, audio, webcam, etc. This is what makes it so valuable in multiple areas of academia, where this paper is expected to be cited.

As you say, it is correct as is. Just trying to focus the attention of the target audience. **I'd like to ensure that readers which have nothing to do with cloud-based infraestructures are attracted from the beginning**. ATM, it is not until the third paragraph starting from the end that the user reads something about the 'host'. Still, it is not clear if 'host' is 'server', 'workstation', or whatever.

Overall, we do know what docker can be used for, but **lots of researchers have never used it and all the references they have are about kubernetes and large scale deployments (money talks)**. One of the strenghs of x11docker is that it is a greatly engineered orthogonal point of view.

---

> * Overall it sounds a bit complex. Maybe some points could be written more simple. I am not sure about this.

I think it will be good to wait until we get one or a few reviews/suggestions from people that is not involved at all. Certainly, non-CS perspectives would be very valuable.

> GUIDock has a similar target as x11docker, but has a quite unpolished code base. I have hardly criticized GUIDock here: [mviereck/x11docker#92 (comment)](https://github.com/mviereck/x11docker/issues/92#issuecomment-460214545)
> Maybe it must be added because JOSS requires to cite similar projects.

IMHO, it is not required at all. You can decide to do it or not. Both options are correct. If you don't feel like doing so (which I would agree with), the argument is easy: *"they haven't been the first with this solution, but the first publishing a paper about this"*. To reference the original source of information, stackoverflow is ok. If you add it because it falls in the group of 'similar projects' with no regard to novelty, many other should be added too:

- Feb 2014: https://github.com/fcwu/docker-ubuntu-vnc-desktop
- Dec 2015: https://github.com/retog/docker-x11-xpra
- Nov 2016: https://github.com/JAremko/docker-x11-bridge
- Apr 2017: https://github.com/lanrat/docker-xpra-html5
- ...

Citing all of these is something I'd expect from a 'regular' publication. But that would be an introduction to a longer article where the implementation details of x11docker are dicussed and compared to other approaches. So, for each feature, you would explain what did exists, how did you put it together, whether you did anything new and how does it compare to others (performance or functionality). However, I think this is out of the scope of this submission. It could be the target for a future update, tho. That's why I asked about it above.

> * The first half of the summary is about Docker in general, the second part about `x11docker` and what it adds in usability/functionality. Maybe this should be divided with subtitles.

I think that the summary is too short to require a subtitle. We should be able to improve it just by reordering the content. For example, wdyt about removing the first sentence and starting with the fifth paragraph? So:

> x11docker allows to run Desktop GUI applications in an isolated environment by running an X display server (...) shared files and GPU/printer/webcam/audio access.
>
> Containerisation in general (...) is completely different.
>
> The properties of containers (...) De Roure (2018)).
>
> Software and required libraries (...) side by side.
>
> The most popular (...) x11docker fills this gap.
>
> x11docker thereby facilitates (...) scholarly review process.
>
> x11docker has (...) MS Windows, too.
