### 2019-03-25 #1349 @whedon
[REVIEW]: x11docker - Run GUI applications in Docker containers
**Submitting author:** @mviereck (<a href="http://orcid.org/0000-0002-4532-4020">Martin Viereck</a>)
**Repository:** <a href="https://github.com/mviereck/x11docker" target ="_blank">https://github.com/mviereck/x11docker</a>
**Version:** v5.6.0

Review based on https://github.com/mviereck/x11docker/commit/56de629
