### 2019-03-25 #1349 @arfon
@yxliang01, @sgyzetrov, @1138-4EB - please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
