### 2019-03-25 mviereck/x11docker#137 @mviereck
JOSS: discussion of paper.md
Let's discuss details of `paper.md` here.

[Link to paper.md](https://github.com/mviereck/x11docker/blob/master/paper.md)
[Link to compiled PDF](https://github.com/openjournals/joss-papers/blob/joss.01346/joss.01346/10.21105.joss.01346.pdf)

Citation from https://github.com/openjournals/joss-reviews/issues/1346:

@mviereck wrote:
Some points I am still not sure about:
 - Overall it sounds a bit complex. Maybe some points could be written more simple. I am not sure about this.
 - @nuest suggested a reference to GUIDock here: https://github.com/mviereck/x11docker/issues/92#issuecomment-459646507
  GUIDock has a similar target as x11docker, but has a quite unpolished code base. I have hardly criticized GUIDock here: https://github.com/mviereck/x11docker/issues/92#issuecomment-460214545
 Maybe it must be added because JOSS requires to cite similar projects.
 - The first half of the summary is about Docker in general, the second part about `x11docker` and what it adds in usability/functionality. Maybe this should be divided with subtitles.
