### 2019-03-25 #1349 @mviereck
Thanks at all of you for your interest and help!

I've opened a ticket for discussion of `paper.md` here: https://github.com/mviereck/x11docker/issues/137
Feel free to open [further tickets ](https://github.com/mviereck/x11docker/issues) for each point to discuss.
