### 2019-03-25 #1349 @1138-4EB
I have finished reviewing the checklist. There are three items which I'd like to explain:

> Conflict of interest

See https://github.com/openjournals/joss-reviews/issues/1346#issuecomment-475979876

> [x] Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Dependencies are clearly-stated, and there are plenty of info messages to tell the user when some are missing. There is no explicit list, tho. I.e., the user must check how a specific package is installed in a host. Nonetheless, I think that an automated package management would unnecessarily complicate this project.

Should I uncheck this item?

> [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

There are no guidelines at all. Words 'contribute'/'contributor' are not found in the repo. This is because it is a single man project, so most of the contributions are not direct code modifications.

However, it is easy to report issues or problems and to get support. The author is very responsive, and most issues are attended and solved within hours or days. PRs are also handled fast.

Should I cross off this item?

---

Regarding https://github.com/openjournals/joss-reviews/issues/1346#issuecomment-476201896, the suggestions still apply, and have been copied to mviereck/x11docker#137. But none of them is a stopper for approval. Those are just suggestions that can opinionatedly improve the quality of the paper.
