### 2019-03-25 #1349 @mviereck
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

> There are no guidelines at all.

I have added some notes about issues, contributing and support in README.md: https://github.com/mviereck/x11docker#contact
