+++ README.md
@@ -344,8 +344,9 @@ For troubleshooting, run `x11docker` or `x11docker-gui` in a terminal.
    
 ## Contact
 Feel free to open a [ticket](https://github.com/mviereck/x11docker/issues) if you have a question or encounter an issue.
+### Issues
 If reporting an issue:
- - Have a look at chapter [Troubleshooting](#troubleshooting)
+ - Have a look at chapter [Troubleshooting](#troubleshooting).
  - Most times it makes sense to store the `--verbose` output (or `~/.cache/x11docker/x11docker.log`) at [pastebin.com](https://pastebin.com/).
 ### Contributing
 If you want to contribute to x11docker, please open a [ticket](https://github.com/mviereck/x11docker/issues) before creating a pull request. Often it is possible to accomplish desired tasks with already available options.
