### 2019-03-25 #1349 @1138-4EB
> I have added some notes about issues, contributing and support in README.md

Review updated accordingly.
