### 2019-03-26 mviereck/x11docker#137 @mviereck
@1138-4EB Thanks for your suggestions!
I've made some careful adjustments in `paper.md`. ([diff](https://github.com/mviereck/x11docker/commit/8f1cc6e4d56db754fe508661acdb643e4fe9ad65))

> The first half of the summary is about Docker in general, the second part about x11docker and what it adds in usability/functionality. Maybe this should be divided with subtitles.

I've added subtitles. Hopefully the pdf converter accepts that.

> I think it can be solved just by replacing 'reduced' with 'minimal'

Done.

> In this case I'm proposing to remove a limiting addition to the sentence, not to add explicit cases: 'x11docker allows to execute Desktop GUI applications in an isolated environment by running an X display server <s>on the host system </s> and providing it to applications in Docker containers'.

I would not like to drop "on the host system". It is a familiar term from virtualisation and I believe that it helps to distinguish host <-> container.

> My main point here is that x11docker is not specially useful in the cloud-based infraestructures where docker is so widespread, but it is a great tool for other not-so-conventional uses of docker,

> Overall, we do know what docker can be used for, but lots of researchers have never used it and all the references they have are about kubernetes and large scale deployments (money talks). One of the strenghs of x11docker is that it is a greatly engineered orthogonal point of view.

You are right. That could be more clear. I believe the division with subtitles already helps. Also I've extended the description of x11docker features a bit.
