### 2019-03-26 #1349 @mviereck
I've made some careful adjustments in `paper.md`. ([diff](https://github.com/mviereck/x11docker/commit/8f1cc6e4d56db754fe508661acdb643e4fe9ad65))
