### 2019-03-26 mviereck/x11docker#137 @sgyzetrov
I am doing a fresh install (on macOS with Ubuntu 14.04 VM in virtual box, also just installed docker in the test environment). And here are some issues I found so far:

N.B. For docker image I am using `x11docker/lxqt`

1. `sudo x11docker x11docker/lxqt` works great for me but `sudo x11docker --desktop x11docker/lxqt` won't work and throw ` xf86OpenConsole: Switching VT failed`. Could related to [xf86OpenConsole: Cannot open virtual console](https://github.com/mviereck/x11docker/issues/5). This is not an issue for JOSS review, and I am sure it has something to do with my test environment, which is an old Ubuntu 14.04 image.
2. For
    > [ ] Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

    `x11docker --desktop --size 320x240 x11docker/lxqt` works great after I got xephyr installed using`sudo apt install xserver-xephyr`, which brings to my next issue (more on documentation side): should the author bring a dedicated script just for installing the dependencies described in:
    > TL;DR: Install xpra Xephyr weston Xwayland xdotool xauth xclip xrandr xdpyinfo on host, or leave it as it is.
3. For the checklist item:
    > [ ] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

    It might be better to extend on the `target audience` part although I am sure the clear-stated `what problems the software is designed to solve` part is ample enough for this item to be checked. But I will leave this unchecked for now.
4. Just a small side note:`curl -fsSL https://raw.githubusercontent.com/mviereck/x11docker/master/x11docker | sudo bash -s -- --update` would probably work, but if I am to start my installation at `x11docker --install` according to README, then `sudo ./x11docker --install` works for me. So I am just wondering if `x11docker --install` in README is not clear enough (there are some less computer-savvy folks out there after all...).
Also, this is from the reviewer checklist item's point of view... :
    > [ ] Does installation proceed as outlined in the documentation?
