### 2019-03-26 #1349 @sgyzetrov
I'm almost finished reviewing the software (and software documentation) part and going to start the paper itself soon. For the software (and software documentation) part, I have raised a few issues in [mviereck/x11docker#137](https://github.com/mviereck/x11docker/issues/137)
