+++ CHANGELOG.md
@@ -8,12 +8,12 @@ Project website: https://github.com/mviereck/x11docker
 
 ## [Unreleased]
 ### Added
- - `--init=s6-overlay`: Support of init system s6 as given by s6-overlay.
+ - `--init=s6-overlay`: Support of init system `s6` as given by `s6-overlay`.
    [(#136)](https://github.com/mviereck/x11docker/issues/136)
 ### Fixed
  - `--dbus-system`: Regression fix for service file check.
  - `--exe`: Regard possible `--` in command. Command before it has been
-   dropped as `DOCKER_RUN_OPTIONS` before.
+   dropped errately as invalid `DOCKER_RUN_OPTIONS` before.
  - `--tty`: Workaround: Set environment variables `LINES` and `COLUMNS` to 
    current terminal size. Only needed without `--interactive`.
    [Docker bug ticket #33794](https://github.com/moby/moby/issues/33794)
