### 2019-03-27 #1349 @sgyzetrov
I have gone through the paper, and it looks good to me. Some wording could be enhanced but since the author's ideas are clear-stated, this is more of a cherry-on-top thing.

Two newly raised issues:
- Reference for jupyter seems broken in pdf submission (e.g. `jupyter notebooks (al., 2018)`), maybe consider using the recommended bibtex entry from [here](https://conference.scipy.org/proceedings/scipy2018/project_jupyter.html), see bib file [here](https://conference.scipy.org/proceedings/scipy2018/bib/project_jupyter.bib). @mviereck please look into it.
- It would be nice to keep the citation with one consistent style (e.g. `Rechert et al., 2017,Emsley & De Roure (2018)` >> `Rechert et al., 2017; Emsley & De Roure, 2018`)

One remaining issue:
https://github.com/mviereck/x11docker/issues/137#issuecomment-476785198
