### 2019-03-27 b8a7c1c
update, explain --install
