### 2019-03-27 mviereck/x11docker#137 @mviereck
@sgyzetrov Thank you!

> 1) sudo x11docker --desktop x11docker/lxqt won't work and throw xf86OpenConsole: Switching VT failed

I did some test runs in a fresh Ubuntu 14.04 VM. I got different error messages, but always some hints in terminal output how to fix this. Didn't you get this message?
```
x11docker WARNING: Your configuration seems not to allow to start
  a second core Xorg server from within X. Option --xorg will probably fail.
  (Per default, only root or console users are allowed to run an Xorg server).

  Possible solutions:
  1.) Install one of nested X servers 'Xephyr', 'Xnest' or 'nxagent'.
      For --gpu support: install 'weston' and 'Xwayland'.
  2.) Switch to console tty1...tty6 with <CTRL><ALT><F1>...<F6>
      and start x11docker there.
  3.) Run x11docker as root.
  4.) Edit file '/etc/X11/Xwrapper.config' and replace line:
          allowed_users=console
      with lines
          allowed_users=anybody
          needs_root_rights=yes
      If the file does not exist already, you can create it.
      On Ubuntu 16.04 and debian 9 you need package xserver-xorg-legacy.
```
I've also found some minor issues with Ubuntu 14.04 that are fixed now in master branch.

-----------------
> 2) should the author bring a dedicated script just for installing the dependencies described in:

I'd like not to do that.
 - Most dependencies are not mandatory but only recommendations.
 - There are many different package managers and different package names
 - Providing different packages for different systems would be the nicest way. But also a lot of work.

>> Ideally these should be handled with an automated package management solution.

As long as it is a recommendation, I won't do this major task yet.

Instead, a possible improvement could be if I provide a table with package names for a number of popular distributions. This makes it easier to find the matching package for each recommended dependency.

-----------------
> 3) who the target audience is?

I consider to add a paragraph in `paper.md`:
> The target audience of `x11docker` in general are users who want to run desktop applications in containers.
Another target audience are developers of desktop applications who need an isolated environment.
In scientific and academic context the target audience are researchers in the field of reproducible science.

-----------------
> 4) So I am just wondering if x11docker --install in README is not clear enough

I've added some explanation in README.md and `x11docker --help`

>  - `x11docker --install` : install x11docker and x11docker-gui from current directory. (Useful to install from an extracted `zip` file or a cloned `git` repository.)
