### 2019-03-27 d35ffba
fix jupyter reference; add target audience #137
