+++ paper.bib
@@ -53,13 +53,16 @@
 	pages = {2396--2398}
 }
 
-@inproceedings{jupyter2018binder,
-  title = {Binder 2.0 - Reproducible, interactive, sharable environments for science at scale},
-  author = {Project Jupyter et al.},
-  year = {2018},
-  booktitle = {Proceedings of the 17th Python in Science Conference},
-  url = {http://conference.scipy.org/proceedings/scipy2018/pdfs/project_jupyter.pdf},
-  doi = {10.25080/Majora-4af1f417-011}
+
+@InProceedings{ project_jupyter-proc-scipy-2018,
+  author    = { {P}roject {J}upyter and {M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing },
+  title     = { {B}inder 2.0 - {R}eproducible, interactive, sharable environments for science at scale },
+  booktitle = { {P}roceedings of the 17th {P}ython in {S}cience {C}onference },
+  url      = {http://conference.scipy.org/proceedings/scipy2018/pdfs/project_jupyter.pdf},
+  pages     = { 113 - 120 },
+  year      = { 2018 },
+  editor    = { {F}atih {A}kici and {D}avid {L}ippa and {D}illon {N}iederhut and {M} {P}acer },
+  doi       = { 10.25080/Majora-4af1f417-011 }
 }
 
 @article{marwick_computational_2017,
