+++ paper.md
@@ -39,7 +39,7 @@ application stacks make it a promising candidate to increase computational repro
 reusability of research analyses [@boettiger_introduction_2015].
 Their use has been demonstrated in various disciplines, such as software engineering research 
 [@cito_using_2016], bioinformatics [@hosny_algorun_2016], and archeology [@marwick_computational_2017], 
-and their preservation is an active field of research [@rechert_preserving_2017;@emsley_framework_2018].
+and their preservation is an active field of research [@rechert_preserving_2017; @emsley_framework_2018].
 
 Software and required libraries can be installed in a Docker image to run software that is difficult 
 to install otherwise. It is possible to run outdated versions, specific versions, or latest development 
