### 2019-03-27 #1349 @mviereck
> Some wording could be enhanced but since the author's ideas are clear-stated, this is more of a cherry-on-top thing.

Suggestions (or a PR) are quite appreciated!

> Reference for jupyter seems broken in pdf submission

fixed.

> It would be nice to keep the citation with one consistent style

done.

> One remaining issue: https://github.com/mviereck/x11docker/issues/137#issuecomment-476785198

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I've added a note about target audience in `paper.md`.

One point is still open:
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

As long as it is a recommendation ("ideally"), I won't do the major task of automated package management.
Instead, I could provide a table with package names for a number of popular distributions. This makes it easier to find the matching package for each recommended dependency.
