### 2019-03-29 #1349 @sgyzetrov
> Instead, I could provide a table with package names for a number of popular distributions. This makes it easier to find the matching package for each recommended dependency.

https://github.com/mviereck/x11docker/commit/d699254fa52513d2774c229d5e5d0b934c698d75 looks good enough to me! Since all issues have been addressed, I move to recommend this paper for acceptance.
