### 2019-03-29 #1349 @arfon
@mviereck - it looks like you have an entry in your bibtex file (`@project_jupyter-proc-scipy-2018`) that isn't being cited in your paper. Please either cite this item directly in the paper or remove it from your bibtex file.
