### 2019-03-29 #1349 @mviereck
> it looks like you have an entry in your bibtex file (@project_jupyter-proc-scipy-2018) that isn't being cited in your paper.

Strange - it is cited in the paper and I didn't change `paper.md` or `paper.bib` since last successful pdf generation. I'll try again.
