### 2019-03-29 #1349 @arfon
> Strange - it is cited in the paper and I didn't change paper.md or paper.bib since last successful pdf generation. I'll try again.

Apologies @mviereck - this is a bug in Whedon which I've now fixed. Please ignore my [previous comment](https://github.com/openjournals/joss-reviews/issues/1349#issuecomment-477890908).
