+++ README.md
@@ -57,9 +57,6 @@ x11docker runs on Linux and (with some setup and limitations) on [MS Windows](#m
    - [Init system](#init-system)
    - [DBus](#dbus)
  - [Dependencies](#dependencies)
-   - [X server dependencies](#x-server-dependencies)
-   - [Option dependencies](#option-dependencies)
-   - [List of all host packages for all possible x11docker options](#list-of-all-host-packages-for-all-possible-x11docker-options)
  - [Security](#security)
    - [Options degrading container isolation](#options-degrading-container-isolation)
    - [Sandbox](#sandbox)
@@ -136,7 +133,10 @@ Other files than `x11docker` script itself are not essential.
 
 
 ## Options
-Description of some commonly used [options](https://github.com/mviereck/x11docker/wiki/x11docker-options-overview).
+Description of some commonly used feature [options](https://github.com/mviereck/x11docker/wiki/x11docker-options-overview).
+ - Some of these options have dependencies on host and/or in image.
+   Compare [wiki: feature dependencies](https://github.com/mviereck/x11docker/wiki/Dependencies#dependencies-of-feature-options).
+
 
 ### Choice of X servers and Wayland compositors
 If no X server option is specified, x11docker automatically chooses one depending on installed [dependencies](#dependencies) and on given or missing options `--desktop`, `--gpu` and `--wayland`.
@@ -165,7 +165,7 @@ Note that x11docker copies files from `/etc/skel` in container to `HOME` if `HOM
  
 ### Hardware acceleration
 Hardware acceleration for OpenGL is possible with option `-g, --gpu`. 
- - This will work out of the box in most cases with open source drivers on host. Otherwise have a look at [Dependencies](#option-dependencies). 
+ - This will work out of the box in most cases with open source drivers on host. Otherwise have a look at [wiki: feature dependencies](https://github.com/mviereck/x11docker/wiki/Dependencies#dependencies-of-feature-options). 
  - Closed source [NVIDIA drivers](https://github.com/mviereck/x11docker/wiki/NVIDIA-driver-support-for-docker-container) need some setup and support less [x11docker X server options](https://github.com/mviereck/x11docker/wiki/X-server-and-Wayland-Options#attributes-of-x-server-and-wayland-options).
  
 ### Clipboard
@@ -175,26 +175,30 @@ Clipboard sharing is possible with option `-c, --clipboard`.
  
 ### Sound
 Sound is possible with options `-p, --pulseaudio` and `--alsa`. 
- - For pulseaudio sound with `--pulseaudio` you need `pulseaudio` on host and in image.
+ - For pulseaudio sound with `--pulseaudio` you need `pulseaudio` on host and `pulseaudio` libraries in image. 
+   Compare [wiki: feature dependencies](https://github.com/mviereck/x11docker/wiki/Dependencies#dependencies-of-feature-options).
  - For ALSA sound with `--alsa` you might need to specify a sound card with e.g. `--alsa=Generic`. Get a list of available sound cards with `aplay -l`.
-   To support virtual ALSA devices like `dmix`, too, the image needs ALSA libraries, e.g. `libasound2` in debian images.
  
 ### Webcam
 Webcams on host can be shared with option `--webcam`.
- - If webcam application in image fails, install `mesa-utils` (debian) or `mesa-demos` (arch) in image. 
+ - If webcam application in image fails, install `--gpu` dependencies in image. 
+   Compare [wiki: feature dependencies](https://github.com/mviereck/x11docker/wiki/Dependencies#dependencies-of-feature-options).
  - `guvcview` needs `--pulseaudio` or `--alsa`.
  - `cheese` and [`gnome-ring`](https://ring.cx/) need `--init=systemd` or `--dbus-system`.
  
 ### Printer
 Printers on host can be provided to container with option `--printer`. 
[…]
