### 2019-03-29 25cad98
reorder content (installation, dependencies). Outsourced to wiki: Installation on MS Windows. #137
