+++ README.md
@@ -13,7 +13,7 @@ Software can be installed in a deployable Docker image with a rudimentary Linux 
 This can help to run or deploy software that is difficult to install on several systems due to dependency issues. It is possible to run outdated versions or latest development versions side by side. 
 Files to work on can be shared between host and container.
 
-x11docker runs on Linux and (with some setup and limitations) on [MS Windows](#msys2-cygwin-and-wsl-on-ms-windows). x11docker is not adapted to run on macOS except in a Linux VM.
+x11docker runs on Linux and (with some setup and limitations) on [MS Windows](#installation-on-ms-windows). x11docker is not adapted to run on macOS except in a Linux VM.
 
 [x11docker wiki](https://github.com/mviereck/x11docker/wiki) provides some how-to's for basic setups without x11docker.
 
@@ -39,10 +39,6 @@ x11docker runs on Linux and (with some setup and limitations) on [MS Windows](#m
 ### Table of contents
  - [GUI for x11docker](#gui-for-x11docker)
  - [Terminal usage](#terminal-usage)
- - [Installation](#installation)
-   - [Installation options](#installation-options)
-   - [Shortest way for first installation](#shortest-way-for-first-installation)
-   - [Minimal installation](#minimal-installation)
  - [Options](#options)
    - [Choice of X servers and Wayland compositors](#choice-of-x-servers-and-wayland-compositors)
    - [Desktop or seamless mode](#desktop-or-seamless-mode)
@@ -56,11 +52,15 @@ x11docker runs on Linux and (with some setup and limitations) on [MS Windows](#m
    - [Wayland](#wayland)
    - [Init system](#init-system)
    - [DBus](#dbus)
- - [Dependencies](#dependencies)
  - [Security](#security)
    - [Options degrading container isolation](#options-degrading-container-isolation)
    - [Sandbox](#sandbox)
- - [MSYS2, Cygwin and WSL on MS Windows](#msys2-cygwin-and-wsl-on-ms-windows)
+ - [Installation](#installation)
+   - [Installation options](#installation-options)
+   - [Shortest way for first installation](#shortest-way-for-first-installation)
+   - [Minimal installation](#minimal-installation)
+   - [Installation on MS Windows](#installation-on-ms-windows)
+ - [Dependencies](#dependencies)
  - [Troubleshooting](#troubleshooting)
  - [Contact](#contact)
    - [Issues](#issues)
@@ -104,34 +104,6 @@ To run only a new empty X server:
  
  
  
-## Installation
-Note that x11docker is just a **bash script** without library dependencies. Basically it is a wrapper for X servers and Docker. To allow advanced usage of x11docker abilities look at chapter [Dependencies](#dependencies).
-### Installation options
-As root you can install, update and remove x11docker on your system:
- - `x11docker --install` : install x11docker and x11docker-gui from current directory. (Useful to install from an extracted `zip` file or a cloned `git` repository.)
- - `x11docker --update` : download and install latest [release](https://github.com/mviereck/x11docker/releases) from github.
- - `x11docker --update-master` : download and install latest master version from github.
- - `x11docker --remove` : remove all files installed by x11docker.
- 
-Copies `x11docker` and `x11docker-gui` to `/usr/bin`. Creates an icon in `/usr/share/icons`. 
-Creates `x11docker.desktop` in `/usr/share/applications`. Copies `README.md`, `CHANGELOG.md` and `LICENSE.txt` to `/usr/share/doc/x11docker`.
-### Shortest way for first installation:
- - For systems using `sudo`:
-   ```
-   curl -fsSL https://raw.githubusercontent.com/mviereck/x11docker/master/x11docker | sudo bash -s -- --update
-   ```
- - Directly as `root`:
-   ```
[…]
