### 2019-03-30 #1349 @yxliang01
@mviereck Few partial comments/questions:

1. In introduction of README, it says "The result is similar to a virtual machine, but needs less resources.". Containers are actually quite different from virtual machine. As such, the important information from this sentence is only that it consumes less resources than virtual machine on the similar tasks which might be more accurate to say so.
2. In the paper, the affiliation is your name. However, I think it is supposed to be an organization name. Since this is a personal project, I feel maybe you can remove the affiliation? Or, did you mean your name is also the name of an organization consist only you?
3. Edit suggestion in Install section of README: `As root you can install, update and remove x11docker on your system` -> `As root you can install, update and remove x11docker to system directories to be available system-wide`. The rationale is that x11docker doesn't need to be installed in system directory to be working fully expectedly, so, essentially putting in your `~/Downloads` folder can also be said as "installed". However, this might be arguable. So, it's acceptable to me if you decide to not modify.

@arfon Few questions for you:
1. I have contributed to this project, also this paper in particular (although not so much, and it was volunteering-based). I have no association with the author @mviereck except matters related to this project(x11docker). I can assure this fact will not affect me from making impartial review to this paper. Am I passing the conflict of interest check?
2. Regarding **Automated tests**, this project doesn't have any test written nor manual examples dedicated for testing the functional correctness of this implementation. So, I think this check is not passing, am I right?

Thanks
