### 2019-03-30 #1349 @mviereck
@yxliang01 Thanks for your suggestions!

1) > the important information from this sentence is only that it consumes less resources than virtual machine on the similar tasks which might be more accurate to say so.

I've reworded to:
"Containers need much less resources than [virtual machines](https://en.wikipedia.org/wiki/Virtual_machine) for similar tasks."
The same sentence has been in `paper.md`. I have reworded it the same way.

2) > I feel maybe you can remove the affiliation?

I agree. Effectively only my name and my location (Germany) make sense.
I've tried pdf generation without affiliations, but @whedon complains.

@arfon How should we handle affiliations? Me/the author is just a private person without a company.

3) > Edit suggestion in Install section of README

Done, also in the part "Minimal installation".
