### 2019-04-02 #1349 @mviereck
> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

To address this point to more then "Run an example image and see if it works" I am currently developing a special security and feature check image `x11docker/check`.

github source: https://github.com/mviereck/dockerfile-x11docker-check
docker hub: https://hub.docker.com/r/x11docker/check
Note in README.md: https://github.com/mviereck/x11docker#security-and-feature-check

The image is in early development stage, but already allows to run some important and impressive checks.
