### 2019-04-07 mviereck/x11docker#137 @1138-4EB
I just want to congratulate @mviereck for the great effort he has put on documenting all the dependencies for multiple platforms: https://github.com/mviereck/x11docker/wiki/dependencies
