### 2019-04-08 #1349 @mviereck
@arfon Could you please have a look?

Three points:
 1. @yxliang01  has a question due to possible conflict of interest. See https://github.com/openjournals/joss-reviews/issues/1349#issuecomment-478289746.
 2. I am a private person. How should affiliations be handled? Compare point 2 of https://github.com/openjournals/joss-reviews/issues/1349#issuecomment-478293243
 3. I consider to move the project away from github to gitlab. I could do that within the next few days. I would like to link the project to the further maintained gitlab repository instead of current github. Is that a problem?
