### 2019-04-10 #1349 @arfon
> I have contributed to this project, also this paper in particular (although not so much, and it was volunteering-based). I have no association with the author @mviereck except matters related to this project(x11docker). I can assure this fact will not affect me making impartial review on this paper. Am I passing the conflict of interest check?

@yxliang01 - thanks for disclosing this. It sounds like you've contributed to this paper as well as the submitted software. I think this _is_ a conflict so I don't think it's appropriate for you to review this submission.

> I am a private person. How should affiliations be handled? Compare point 2 of #1349 (comment)

You can just have an affiliation of 'None' but you have to make an entry in the `paper.md` YAML called `none` e.g.:

```
affiliations:
  - name: None
     index: 1
```
