### 2019-04-10 #1349 @1138-4EB
> @yxliang01 - thanks for disclosing this. It sounds like you've contributed to this paper as well as the submitted software. I think this _is_ a conflict so I don't think it's appropriate for you to review this submission.

If this is the case, none of us should review it:

- https://github.com/openjournals/joss-reviews/issues/1349#issuecomment-476301706
- https://github.com/openjournals/joss-reviews/issues/1349#issuecomment-476270078
