### 2019-04-10 #1349 @arfon
> If this is the case, none of us should review it:

Not necessarily. I had interpreted @yxliang01's comment that they had contributed to both the software, and the JOSS paper being reviewed here. Perhaps I misunderstood your comment @yxliang01?
