### 2019-04-10 #1349 @mviereck
@arfon
You can see the contributions of @yxliang01 to the paper here:
One small PR for paper.md: https://github.com/mviereck/x11docker/pull/98
A few comments in the pre-submission discussion: https://github.com/mviereck/x11docker/issues/92

Only one comment of @sgyzetrov here before submission: https://github.com/mviereck/x11docker/issues/92#issuecomment-459775959

I don't remember issue tickets of @yxliang01 and @sgyzetrov yet.
Their contributions to `paper.md` could have happened in this submission thread as well.
I would consider their contributions to be harmless.

More previous contribution has been by @1138-4EB who helped me to intensively check out x11docker on MS Windows along the past ~12 months. (No direct code contribution, though.)
He has stated this here: https://github.com/openjournals/joss-reviews/issues/1346#issuecomment-475979876

All three reviewers have known x11docker before submission as @yxliang01 noted here: https://github.com/openjournals/joss-reviews/issues/1349#issuecomment-476270078
