### 2019-04-10 #1349 @yxliang01
@arfon Yes. To clarify, I have never contributed to the software and the contribution was on the paper only which I have done the same in a form of reviews in this thread also. In that sense, I am a volunteer paper editor for this paper and not an author. Since I think we have mutual understanding now and your suggestion, I will tick the *conflict of interest* and continue my review with further potential edit suggestions.

Thanks for the clarification.
