+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Project website: https://github.com/mviereck/x11docker
 
 ## [Unreleased]
+### Added
+ - Experimental support of `--runtime=kata-runtime` in custom docker run 
+   options.
+   [(#138)](https://github.com/mviereck/x11docker/issues/138)
 ### Fixed
  - `--init=runit`: Create DBus service files if missing.
  - `docker commit`: Throw error if running a recursive image created from 
