### 2019-04-14 e97954e
new option: --share. replaces --sharedir. --runtime=kata-runtime: some finetuning, catch unsupported options #138
