### 2019-04-17 94e2c6c
set affiliations to "None"
