+++ CHANGELOG.md
@@ -17,9 +17,10 @@ Project website: https://github.com/mviereck/x11docker
  - `--no-xhost`: No replacement.
  - `--sys-admin`: No replacement.
 ### Fixed
- - `--init=runit`: Create runit service for DBus if missing.
- - `docker commit`: Throw error if running a recursive image created from 
-   an x11docker container with `docker commit`.
+ - `--dbus-system`: Fixed startup failure with user switching.
+ - `--init=runit|openrc`: Create service to start DBus if missing.
+ - `docker commit`: Throw error if running a recursive image command
+   created with `docker commit` from an x11docker container.
    [(#146)](https://github.com/mviereck/x11docker/issues/146)
 
 ## [5.5.2](https://github.com/mviereck/x11docker/releases/tag/v5.5.2) - 2019-04-08
