### 2019-04-25 #1349 @yxliang01
@mviereck Sorry about the delay. I have read through the paper and have the following doubts:

1.  [Typo?] "dependency packaging" and "minimal requirements of system environment (i.e. only the container runtime)" are not properties of containers. "dependency packaging" is instead something can be achieved via containerization.
2. [Mirror] You mentioned that virtualization and containers are technically ***completely*** different. Arguably, both KVM and containers are implemented in the Linux kernel. In that sense, they are similar in the sense that the support in the kernel.
3. [Unverified Performance Claim] In the paper, it says "rather slow interaction due to a lot of network data transfer". Any support for this? At least, to the best of my knowledge, XPRA supports UNIX Socket connection if you mean IP protocol connections by "network data transfer".

Besides these doubts, I will fine with other parts of this submission.

@arfon
I have just checked the "Automated tests" check because there are detailed documentations to guide verifying the functionality of the software that I think satisfying the "manual steps" option of the check. However, `x11docker` doesn't have any test. Please advise if I misunderstood the passing criteria of the check.
