### 2019-04-25 #1349 @yxliang01
@mviereck I have just also suggested edits in the PR https://github.com/mviereck/x11docker/pull/149 . Some of the suggestions are related to the doubts mentioned in comment https://github.com/openjournals/joss-reviews/issues/1349#issuecomment-486647502 and also mirror changes.

**Note that the created PR has no way to block the review process of this submission and should be considered independent from this submission process**.
