### 2019-04-25 #1349 @yxliang01
@arfon The generated PDF has incorrect reference in the footer. Maybe you would want to check the PDF compilation?
