### 2019-04-25 #1349 @arfon
> @arfon The generated PDF has incorrect reference in the footer. Maybe you would want to check the PDF compilation?

This is because of `paper.yml` file. The name is currently: `name: Martin Viereck, Germany`
