### 2019-04-28 #1349 @mviereck
> maybe you want to modify the name section to only contain your name without the country?

Of course, done in https://github.com/mviereck/x11docker/commit/b109cc0bd0306301941589598accc404671d9e54

I've also changed `affiliations: None` to `affiliations: No affiliations`. I believe that looks better in the title.

> [Unverified Performance Claim] In the paper, it says "rather slow interaction due to a lot of network data transfer". Any support for this? At least, to the best of my knowledge, XPRA supports UNIX Socket connection if you mean IP protocol connections by "network data transfer".

Setups with VNC or HTML5 are terribly slow in general, I don't think I need to prove that. However, you are right that an xpra setup with shared unix sockets is significantly better than that.
I am currently not sure if/how to reword this. I'll think about it.

> @arfon
I have just checked the "Automated tests" check because there are detailed documentations to guide verifying the functionality of the software that I think satisfying the "manual steps" option of the check. However, x11docker doesn't have any test. Please advise if I misunderstood the passing criteria of the check.

> [ ] Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

I'd say that the manual steps in image `x11docker/check` fit this criteria. However, it is not an automated test on itself and needs some interaction.
I'll improve the description of the manual steps in the GUI of `x11docker/check`.
