### 2019-04-28 #1349 @yxliang01
@mviereck Indeed, for comparing with HTML5, intuitively, it would be slower. That can be treated as "common sense" I believe. However, for the case of VNC, I am unsure. The best way to prove this would be having a benchmark result if feasible(I guess not too straightforward and time-consuming?). If infeasible and you are willing, you can also remove this performance claim?
