### 2019-04-28 #1349 @mviereck
> If infeasible and you are willing, you can also remove this performance claim?

Ok, I have just dropped the performance claim. It is not essential overall.
The next pdf generation won't contain the sentence.
