### 2019-04-28 #1349 @arfon
> @arfon
> I have just checked the "Automated tests" check because there are detailed documentations to guide verifying the functionality of the software that I think satisfying the "manual steps" option of the check. However, `x11docker` doesn't have any test. Please advise if I misunderstood the passing criteria of the check.

👍 this is OK.
