+++ README.md
@@ -52,6 +52,7 @@ x11docker runs on Linux and (with some setup and limitations) on [MS Windows](#i
    - [Wayland](#wayland)
    - [Init system](#init-system)
    - [DBus](#dbus)
+   - [Container runtime](#container-runtime)
  - [Security](#security)
    - [Options degrading container isolation](#options-degrading-container-isolation)
    - [Sandbox](#sandbox)
@@ -200,7 +201,39 @@ Some desktop environments and applications need a running DBus daemon and/or DBu
  - use `--sharedir /run/dbus/system_bus_socket` to share host DBus system socket.
  - DBus will be started automatically with [init systems](#Init-system) `systemd`, `openrc`, `runit` and `sysvinit` (option `--init`).
 
+### Container runtime
+It is possible to run containers with different backends following the [OCI runtime specification](https://github.com/opencontainers/runtime-spec). Docker's default is `runc`. You can specify another one with option `--runtime=RUNTIME`.
+Container runtimes known and supported by x11docker are:
+ - `runc`: Docker default.
+ - [`kata-runtime`](https://katacontainers.io/): Sets up a virtual machine with its own Linux kernel to run the container. `kata` aims to combine the security advantages of containers and virtual machines.
+   - Some x11docker options are not possible with `--runtime=kata-runtime`. Most important: `--hostdisplay`, `--gpu`, `--printer`, `--webcam` and all Wayland related options.
+ - `nvidia`: Specialized fork of `runc` to support `nvidia/nvidia-docker` images.
+ - [`crun`](https://github.com/giuseppe/crun): Fast and lightweight alternative to `runc` with same functionality.
  
+Note that currently option `--runtime` is in master branch only. It will be part of the next x11docker release.
+
+Possible runtime configuration in `/etc/docker/daemon.json`:
+```
+{
+  "default-runtime": "runc",
+  "runtimes": {
+    "kata-runtime": {
+      "path": "/opt/kata/bin/kata-runtime",
+      "runtimeArgs": [
+        "--kata-config /opt/kata/share/defaults/kata-containers/configuration.toml"
+      ]
+    },
+    "crun": {
+      "path": "/usr/local/bin/crun",
+      "runtimeArgs": []
+    },
+    "nvidia": {
+      "path": "nvidia-container-runtime",
+      "runtimeArgs": []
+    }
+  }
+}
+```
  
 ## Security 
 Scope of x11docker is to run containerized GUI applications while preserving and improving container isolation.
@@ -224,14 +257,19 @@ I am not aware of an escape from a container without an additional isolation deg
 However, x11docker follows the [principle of least privilege](https://en.wikipedia.org/wiki/Principle_of_least_privilege). 
 Docker containers should not have capabilities or privileges that they don't need for their job.
 
+Docker's default runtime `runc` uses Linux namespaces to isolate container applications, but shares the kernel from host. 
+If you are concerned about container access to host kernel, consider to use [container runtime](#container-runtime) `kata-runtime` instead.
+
 _Weaknesses:_
  - Possible SELinux restrictions are degraded for x11docker containers with docker run option `--security-opt label=type:container_runtime_t` to allow access to new X unix socket. 
    A more restrictive solution is desirable.
    Compare: [SELinux and Docker: allow access to X unix socket in /tmp/.X11-unix](https://unix.stackexchange.com/questions/386767/selinux-and-docker-allow-access-to-x-unix-socket-in-tmp-x11-unix)
- - User namespace remapping is disabled to allow options `--home` and `--homedir` without file ownership issues. (Though, this is less an issue because x11docker already avoids root in container). 
-   Exception: User namespace remapping is not disabled for `--user=RETAIN`.
[…]
