### 2019-04-28 #1349 @mviereck
> Please let me know when you are done amending.

Ok, I am ready. :-)
In `paper.md` I just fixed one typo (`Container` to `container` in a subtitle.)

@arfon:
I would like to change the submitted x11docker version from `5.5.1` to `5.6.0` (upcoming release, can be released right now). Can that be done?

Changes important for this review have been done in `README.md` and in the output of `x11docker --help` (and of course in `paper.md` and `paper.bib`).

Other changes touch rarely needed edge cases, most of them are bug fixes for several init systems in container and a [new option `--runtime`](https://github.com/mviereck/x11docker#container-runtime).

The changes are documented in [CHANGELOG.md](https://github.com/mviereck/x11docker/blob/master/CHANGELOG.md).
