### 2019-04-29 #1349 @mviereck
> are you finished with your updates based on feedback from the reviews here?

Yes.
I would only need to set the version number and publish release `v5.6.0`, can be done within minutes.
