### 2019-04-29 #1349 @arfon
OK great!

@yxliang01, @sgyzetrov, @1138-4EB - could you confirm you're happy with the updates from @mviereck? If so, I think we're ready to accept this submission into JOSS.
