### 2019-04-29 8e79846
--share: support device files in /dev
