### 2019-05-02 #1349 @mviereck
I have archived x11docker at Zenodo: https://zenodo.org/record/2658749

> Viereck, Martin. (2019, May 2). x11docker: Run GUI applications in Docker containers (Version v5.6.0). Zenodo. http://doi.org/10.5281/zenodo.2658749

The archived files are the identical to x11docker release v5.6.0 at github: https://github.com/mviereck/x11docker/releases/tag/v5.6.0
